LEM tag own file and tests
